Fixing codacy, formatting issues and addressing review comments
