+++ runtime/src/main/java/org/corfudb/runtime/view/RuntimeLayout.java
@@ -14,7 +14,6 @@ import org.corfudb.runtime.clients.SequencerClient;
 import org.corfudb.runtime.exceptions.QuorumUnreachableException;
 import org.corfudb.runtime.exceptions.WrongEpochException;
 import org.corfudb.runtime.exceptions.unrecoverable.UnrecoverableCorfuError;
-import org.corfudb.runtime.object.transactions.TransactionalContext;
 import org.corfudb.runtime.view.Layout.LayoutSegment;
 
 import javax.annotation.Nonnull;
